requirements.txt
discord.py==2.6.4
discord-ext-voice_recv==0.5.2a179
numpy==2.4.2
python-dotenv==1.2.1
PyNaCl==1.5.0
scipy==1.17.0
vosk==0.3.45
